scripts/requirements.txt
duckduckgo-search
Pillow
requests
